backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary>=2.9.10
cloudinary==1.36.0
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
firebase-admin==6.4.0
Pillow>=10.2.0
django-filter==23.5

